Split out fast_decompose
