+++ core/src/algorithms/mod.rs
@@ -1,3 +1,5 @@
 mod btree_merge;
+mod fast_decompose;
 
 pub(crate) use btree_merge::btree_merge;
+pub(crate) use fast_decompose::fast_decompose;
