+++ core/src/commands/mod.rs
@@ -1,7 +1,7 @@
 mod factorize;
 mod search;
 
-pub(crate) use factorize::fast_decompose;
 pub use factorize::{factorize, Factors};
-pub(crate) use search::search_impl;
 pub use search::{search, SearchResult};
+
+pub(crate) use search::search_impl;
